Se genera filtro de búsqueda dinámico para tareas, incluye dto. Se eliminan las clases validadoras, y se reemplazan por método privado 'findOneOrFail' dentro de cada service. Método post para Estado y Prioridad con el fin de facilitar pruebas en el desarrollo del frontEnd.

+++ backend/src/proyectos/proyectos.service.ts
@@ -1,16 +1,20 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateProyectoDto } from './dto/create-proyecto.dto';
 import { UpdateProyectoDto } from './dto/update-proyecto.dto';
 import { Proyecto } from './entities/proyecto.entity';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UsuariosService } from 'src/usuarios/usuarios.service';
-import { timeStamp } from 'console';
-import { ValidarProyecto } from './validarProyecto';
+import { Tarea } from 'src/tareas/entities/tarea.entity';
 
 @Injectable()
 export class ProyectosService {
-  constructor(@InjectRepository(Proyecto) private readonly proyectoRepo: Repository<Proyecto>, private readonly validarProyecto: ValidarProyecto, private readonly usuarioService: UsuariosService){ }
+  constructor(@InjectRepository(Proyecto) 
+  private readonly proyectoRepo: Repository<Proyecto>, 
+  @InjectRepository(Tarea)
+  private readonly tareaRepo: Repository<Tarea>,
+  private readonly usuarioService: UsuariosService  
+  ){}
   
   async create(createProyectoDto: CreateProyectoDto) {
 
@@ -25,7 +29,7 @@ export class ProyectosService {
   }
 
   async findOne(id: number) {
-    const proyecto = await this.validarProyecto.validarIdProyecto(id)
+    const proyecto = await this.findOneOrFail(id)
     return proyecto;
   }
 
@@ -35,9 +39,22 @@ export class ProyectosService {
     return this.proyectoRepo.save(proyectoActualizado);
   }
 
-  async remove(id: number) {
-    const proyecto =  await this.validarProyecto.validarIdProyecto(id)
+  async remove(id: number, force = false) {
+    const proyecto =  await this.findOneOrFail(id);
+    const tareasCount = await this.tareaRepo.count({ where: { proyecto: { id } } });
+
+    if (tareasCount > 0 && !force){
+      throw new ConflictException(
+        `El proyecto tiene ${tareasCount} tarea(s) asociada(s). Requiere confirmación.`,
+      )
+    }
     await this.proyectoRepo.remove(proyecto)
     return { message: 'Se eliminó el proyecto'};
+  }
+
+  private async findOneOrFail(id: number): Promise <Proyecto> {
+    const proyecto = await this.proyectoRepo.findOne({ where: { id } });
+    if (!proyecto) throw new NotFoundException('No se encontró el proyecto.');
+    return proyecto;
   }
 }
